+++ packages/react/components/popover/src/PopoverContent/PopoverContent.tsx
@@ -1,56 +1,69 @@
 import * as React from 'react';
-import { DismissButton, useModal, useOverlay } from '@react-aria/overlays';
+import { DismissButton, usePopover } from '@react-aria/overlays';
 import { cx } from '@project44-manifest/react-styles';
 import type { ForwardRefComponent } from '@project44-manifest/react-types';
 import { mergeProps } from '@project44-manifest/react-utils';
 import { useMergedRef } from '@project44-manifest/use-merged-ref';
-import { usePopoverContext } from '../Popover.context';
 import { StyledPopoverContent } from './PopoverContent.styles';
 import type { PopoverContentElement, PopoverContentProps } from './PopoverContent.types';
 
 export const PopoverContent = React.forwardRef((props, forwardedRef) => {
-  const { as, children, className: classNameProp, css, ...other } = props;
-
   const {
-    isDismissable,
+    as,
+    children,
+    className: classNameProp,
+    css,
+    isKeyboardDismissDisabled = false,
     isNonModal,
-    isKeyboardDismissDisabled,
+    maxHeight,
+    offset = 8,
     onClose,
-    overlayRef,
-    overlayProps,
+    placement = 'bottom',
+    scrollRef,
+    shouldFlip = true,
+    shouldUpdatePosition = true,
     state,
-    shouldCloseOnBlur,
-    shouldCloseOnInteractOutside,
-  } = usePopoverContext();
+    triggerRef,
+    ...other
+  } = props;
 
-  const mergedRef = useMergedRef(overlayRef, forwardedRef);
-
-  const { overlayProps: popoverProps } = useOverlay(
+  const popoverRef = React.useRef<HTMLDivElement>(null);
+  const { popoverProps } = usePopover(
     {
-      isDismissable: isDismissable && state.isOpen,
       isKeyboardDismissDisabled,
-      isOpen: state.isOpen,
-      onClose,
-      shouldCloseOnBlur,
-      shouldCloseOnInteractOutside,
+      isNonModal,
+      maxHeight,
+      offset,
+      placement,
+      popoverRef,
+      scrollRef,
+      shouldFlip,
+      shouldUpdatePosition,
+      triggerRef,
     },
-    overlayRef,
+    state,
   );
 
-  const { modalProps } = useModal({ isDisabled: isNonModal });
+  const handleClose = React.useCallback(() => {
+    onClose?.();
+
+    state?.close();
+  }, [onClose, state]);
+
+  const mergedRef = useMergedRef(popoverRef, forwardedRef);
 
   return (
     <StyledPopoverContent
-      {...mergeProps(popoverProps, modalProps, overlayProps, other)}
+      {...mergeProps(popoverProps, other)}
       ref={mergedRef}
       as={as}
       className={cx('manifest-popover', classNameProp)}
       css={css}
       role="presentation"
     >
-      {!isNonModal && <DismissButton onDismiss={onClose} />}
+      {!isNonModal && <DismissButton onDismiss={handleClose} />}
       {children}
-      <DismissButton onDismiss={onClose} />
+      <DismissButton onDismiss={handleClose} />
     </StyledPopoverContent>
   );
 }) as ForwardRefComponent<PopoverContentElement, PopoverContentProps>;
